backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
firebase-admin==6.2.0
google-cloud-aiplatform==1.35.0
googlemaps==4.10.0
pydantic==2.4.2
pydantic-settings==2.0.3
requests==2.31.0
typing-extensions==4.8.0
pytest==7.4.2
httpx==0.25.1
email-validator==2.1.0.post1
